Set Flaot  Result in TextField

+++ NewCalculator/NewCalculator/ViewController.swift
@@ -31,30 +31,13 @@ class ViewController: UIViewController {
     var selectNumber = Int()
     
     var method = Int()
-    var totalRunning = Int()
+    var totalRunning = Float()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         var frameRect:  CGRect = screenText.frame
         frameRect.size.height = 100
         screenText.frame = frameRect
-//        b1.layer.cornerRadius = 25.0
-//        b2.layer.cornerRadius = 25.0
-//        b3.layer.cornerRadius = 25.0
-//        b4.layer.cornerRadius = 25.0
-//        b5.layer.cornerRadius = 25.0
-//        b6.layer.cornerRadius = 25.0
-//        b7.layer.cornerRadius = 25.0
-//        b8.layer.cornerRadius = 25.0
-//        b9.layer.cornerRadius = 25.0
-//        b0.layer.cornerRadius = 25.0
-//        bAC.layer.cornerRadius = 25.0
-//        bAdd.layer.cornerRadius = 25.0
-//        bMul.layer.cornerRadius = 25.0
-//        bSub.layer.cornerRadius = 25.0
-//        bDiv.layer.cornerRadius = 25.0
-//        bEql.layer.cornerRadius = 25.0
-        // Do any additional setup after loading the view, typically from a nib.
     }
 
     override func didReceiveMemoryWarning() {
@@ -126,23 +109,23 @@ class ViewController: UIViewController {
     @IBAction func actionResult(sender: AnyObject) {
         if(totalRunning == 0)
         {
-            totalRunning = selectNumber
+            totalRunning = Float(selectNumber)
         }
         else
         {
             switch(method)
             {
             case 1:
-                totalRunning = totalRunning * selectNumber
+                totalRunning = totalRunning * Float(selectNumber)
                 break;
             case 2:
-                totalRunning = totalRunning / selectNumber
+                totalRunning = totalRunning / Float(selectNumber)
                 break;
             case 3:
-                totalRunning = totalRunning+selectNumber
+                totalRunning = totalRunning+Float(selectNumber)
                 break;
             case 4:
-                totalRunning = totalRunning-selectNumber
+                totalRunning = totalRunning-Float(selectNumber)
                 break;
             default :
                 break;
@@ -162,20 +145,20 @@ class ViewController: UIViewController {
     
     @IBAction func actionDevide(sender: AnyObject) {
         if(totalRunning == 0) {
-            totalRunning = selectNumber
+            totalRunning = Float(selectNumber)
         } else {
             switch(method) {
             case 1:
-                totalRunning = totalRunning * selectNumber
+                totalRunning = totalRunning * Float(selectNumber)
                 break;
             case 2:
-                totalRunning = totalRunning/selectNumber
+                totalRunning = totalRunning/Float(selectNumber)
                 break;
             case 3:
-                totalRunning = totalRunning + selectNumber
+                totalRunning = totalRunning + Float(selectNumber)
                 break;
             case 4:
-                totalRunning = totalRunning - selectNumber
+                totalRunning = totalRunning - Float(selectNumber)
                 break;
             default :
                 break;
@@ -188,23 +171,23 @@ class ViewController: UIViewController {
     @IBAction func actionMultiply(sender: AnyObject) {
         if(totalRunning == 0)
         {
-            totalRunning = selectNumber
+            totalRunning = Float(selectNumber)
         }
         else
         {
             switch(method)
             {
             case 1:
-                totalRunning = totalRunning*selectNumber
+                totalRunning = totalRunning*Float(selectNumber)
                 break;
             case 2:
-                totalRunning = totalRunning/selectNumber
+                totalRunning = totalRunning/Float(selectNumber)
                 break;
             case 3:
-                totalRunning = totalRunning+selectNumber
+                totalRunning = totalRunning+Float(selectNumber)
                 break;
             case 4:
-                totalRunning = totalRunning-selectNumber
+                totalRunning = totalRunning-Float(selectNumber)
                 break;
             default :
                 break;
@@ -218,23 +201,23 @@ class ViewController: UIViewController {
     @IBAction func actionSum(sender: AnyObject) {
         if(totalRunning == 0)
         {
-            totalRunning = selectNumber
+            totalRunning = Float(selectNumber)
         }
         else
         {
             switch(method)
             {
             case 1:
-                totalRunning = totalRunning*selectNumber
+                totalRunning = totalRunning*Float(selectNumber)
                 break;
             case 2:
-                totalRunning = totalRunning/selectNumber
+                totalRunning = totalRunning/Float(selectNumber)
                 break;
             case 3:
-                totalRunning = totalRunning+selectNumber
+                totalRunning = totalRunning+Float(selectNumber)
                 break;
             case 4:
-                totalRunning = totalRunning-selectNumber
+                totalRunning = totalRunning-Float(selectNumber)
                 break;
             default :
                 break;
@@ -247,23 +230,23 @@ class ViewController: UIViewController {
     @IBAction func actionMinus(sender: AnyObject) {
         if(totalRunning == 0)
         {
-            totalRunning = selectNumber
+            totalRunning = Float(selectNumber)
         }
         else
         {
             switch(method)
             {
             case 1:
-                totalRunning = totalRunning*selectNumber
+                totalRunning = totalRunning*Float(selectNumber)
                 break;
             case 2:
-                totalRunning = totalRunning/selectNumber
+                totalRunning = totalRunning/Float(selectNumber)
                 break;
             case 3:
-                totalRunning = totalRunning+selectNumber
+                totalRunning = totalRunning+Float(selectNumber)
                 break;
             case 4:
-                totalRunning = totalRunning-selectNumber
+                totalRunning = totalRunning-Float(selectNumber)
                 break;
             default :
                 break;
